Function faucon::code::upload_to_imem

pub fn upload_to_imem(
    cpu: &mut Cpu, 
    address: u16, 
    vaddress: u32, 
    binary: &[u8]
) -> Result<(), ()>

Uploads a Falcon binary that was obtained from read_falcon_binary into the code segment of the processor.

Returns an error if the binary is too large to fit into the Falcon code segment.