Struct faucon::config::Falcon

pub struct Falcon {
    pub version: usize,
    imem_size: u32,
    pub dmem_size: u32,
    clock_freq: f32,
    pub dma: Dma,
}

Generic Falcon MCU configuration options.

Fields
version: usize

The version of the Falcon MCU to be emulated.

imem_size: u32

The size of the Falcon code space.

dmem_size: u32

The size of the Falcon data space.

clock_freq: f32

The clock frequency of the Falcon, measured in cycles/second.

dma: Dma

The Falcon DMA engine configuration.

Implementations

impl Falcon

pub fn get_imem_size(&self) -> u32

Gets the correct size of the Falcon code segment.

Falcon revisions prior to v3 use a linear code segment with byte-oriented addressing, whereas more recent Falcons use paging with 0x100 byte pages. Depending on the version, this function returns the total size of the segment in bytes.

pub fn cycle_duration(&self) -> f32

Gets the duration of a single clock cycle of the CPU.